requirements.txt

python-telegram-bot>=20.0
httpx
aiohttp
numpy
python-dotenv
requests